moved contextmanager to utils
